services/py/pyproject.toml
[project]
name = "geo-platform"
version = "0.1.0"
description = "GEO Platform — AI Visibility Analysis Engine"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    # --- API Layer ---
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",

    # --- LLM Interface ---
    "litellm>=1.60.0",
    "instructor>=1.7.0",

    # --- Agent Orchestration ---
    "langgraph>=0.3.0",

    # --- Search APIs ---
    "tavily-python>=0.5.0",

    # --- Web Crawling ---
    "httpx>=0.28.0",
    "beautifulsoup4>=4.12.0",
    "readability-lxml>=0.8.0",

    # --- NLP & Sentiment ---
    "spacy>=3.8.0",

    # --- Data & Validation ---
    "polars>=1.20.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",

    # --- Config ---
    "pyyaml>=6.0.0",

    # --- Logging ---
    "structlog>=24.0.0",

    # --- Observability ---
    "langfuse>=2.0.0",

    # --- Security ---
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
# JS-heavy page crawling (Phase 2 fallback)
crawl = ["crawl4ai>=0.4.0"]

# Dev/test dependencies
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.35.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
